+++ src/sim/AveragingTimerCallback.cpp
@@ -26,7 +26,7 @@ void AveragingTimerCallback::run(uint64_t ticks,
   row.m_time = simTime;
 
   m_kernel->calculateAverages();
-  for (int i = 0; i < m_avgVols.size(); i++) {
+  for (size_t i = 0; i < m_avgVols.size(); i++) {
     VoxelVolume avgVol = m_avgVols.at(i);
     LatticeAverage lAvg = m_kernel->getAverage(avgVol, deltaTicks);
